Deduplicate tests
